Add: trajectory execution with new structure
